Fix:401 refactored electionguardtools
