+++ src/electionguardtools/factories/__init__.py
@@ -1,0 +1,2 @@
+from .ballot_factory import *
+from .election_factory import *
