+++ src/electionguardtools/factories/ballot_factory.py
