+++ src/electionguardtools/strategies/__init__.py
@@ -1,0 +1,4 @@
+...
+from .election import *
+from .elgamal import *
+from .group import *
