+++ src/electionguardtools/strategies/group.py
