+++ tests/property/test_chaum_pedersen.py
@@ -19,8 +19,8 @@ from electionguard.elgamal import (
 )
 from electionguard.group import ElementModQ, TWO_MOD_Q, ONE_MOD_Q, int_to_p, TWO_MOD_P
 from electionguard.utils import get_optional
-from electionguardtest.elgamal import elgamal_keypairs
-from electionguardtest.group import elements_mod_q_no_zero, elements_mod_q
+from electionguardtools.strategies.elgamal import elgamal_keypairs
+from electionguardtools.strategies.group import elements_mod_q_no_zero, elements_mod_q
 
 
 class TestDisjunctiveChaumPedersen(TestCase):
