+++ tests/property/test_decryption_mediator.py
@@ -34,12 +34,12 @@ from electionguard.tally import (
 )
 from electionguard.utils import get_optional
 
-import electionguardtest.ballot_factory as BallotFactory
-import electionguardtest.election_factory as ElectionFactory
-from electionguardtest.election import election_descriptions, plaintext_voted_ballots
-from electionguardtest.decryption_helper import DecryptionHelper
-from electionguardtest.key_ceremony_helper import KeyCeremonyHelper
-from electionguardtest.tally import accumulate_plaintext_ballots
+import electionguardtools.factories.ballot_factory as BallotFactory
+import electionguardtools.factories.election_factory as ElectionFactory
+from electionguardtools.strategies.election import election_descriptions, plaintext_voted_ballots
+from electionguardtools.helpers.tally_ceremony_orchestrator import DecryptionHelper
+from electionguardtools.helpers.key_ceremony_orchestrator import KeyCeremonyHelper
+from electionguardtools.helpers.tally_accumulate import accumulate_plaintext_ballots
 
 election_factory = ElectionFactory.ElectionFactory()
 ballot_factory = BallotFactory.BallotFactory()
