+++ tests/unit/test_manifest.py
@@ -9,8 +9,8 @@ from electionguard.manifest import (
     VoteVariationType,
 )
 from electionguard.serializable import read_json
-import electionguardtest.election_factory as ElectionFactory
-import electionguardtest.ballot_factory as BallotFactory
+import electionguardtools.factories.election_factory as ElectionFactory
+import electionguardtools.factories.ballot_factory as BallotFactory
 
 
 election_factory = ElectionFactory.ElectionFactory()
